added scripts
this is the first adition to version control of this project. in this files contains several features: CerSetup sets the car up by loading the required meshes, it gives colliders, rigidbodys and others identification data. BasicMovement gives the car the basic movility of the car, that is, direction and traction. explosiveController and proyectileController are used to control bullets and explosives. Cooldown is a single class to control timing in several ways. CameraSetting and CameraController, set and move camera(s) acording to players setup. The mod folder contains modules that provide several aditional behaviors to the car, they are name descriptive. InputConfig is an editor script used to setup the inputs in the unity project. ButtonManager controls the players setup in a pre game menu. SelfSlider provides a hovering health bar above the car.