multiplayer is now working for 2 players
several changes in ordet to setup ultiple players, now the canvas is cloned for each player, cars are added and setted up by script for each player, was nessesary to change all the mods to make universal the setup with dependences, like the layer management, the materials and gui components. The basic logics are scalable for unlimited players, but theres still necessary to manualy set some variables in order to add more players, this can be fixed tho.